requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
Django==2.2.10
django-cors-headers==3.2.1
django-safedelete==0.5.4
graphene-django==2.15.0
isort==4.3.21
jsonfield==2.0.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.1.0
promise==2.3
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
wrapt==1.11.2
django-environ==0.4.5
psycopg2-binary==2.8.6
whitenoise==5.2.0
django-graphql-jwt==0.3.4
backoff==1.11.1
requests==2.26.0
django-filter==2.4.0
celery==5.0.5
redis==3.5.3
